Bug fix for `dpdraw`.
The attribute `alpha` for the `config` struct was an int, where it
should have been a double. This commit corrects this error in
`dpdraw.h`. Also, adding `exit(1)` to `parse_args` in `dpdraw.c` to exit
if the alpha argument is invalid.

+++ src/dpdraw.h
@@ -37,7 +37,7 @@
 typedef struct config_ {
     int num_elements;
     int reps;
-    int alpha;
+    double alpha;
     int seed;
 } config;
